demrepofdave/al5int/resize_windows: common, allegro_4, and allegro_5 directories rearranged in host_abstraction_layer directory to keep all host code separate from electron logic
