+++ Makefile.am
@@ -1,6 +1,6 @@
 #SUBDIRS = src
 
-# SUBDIRS = src src/allegro_4
+# SUBDIRS = src src/host_abstraction_layer/allegro_4/
 
 # Makefile.am for Elkulator
 
@@ -23,37 +23,37 @@ elkulator_LDADD = $(allegro_LIBS) -lz -lopenal -lalut -lz
 endif
 
 if ENABLE_ALLEGRO4
-elkulator_hal_src = src/allegro_4/2xsai.c \
-                    src/allegro_4/adc.c  \
-                    src/allegro_4/palfilt.c \
-                    src/allegro_4/scale2x.c \
-                    src/allegro_4/soundopenal.c \
-                    src/allegro_4/firstbyte.c \
-                    src/allegro_4/linux-gui.c \
-                    src/allegro_4/linux-keydefine.c \
-                    src/allegro_4/video.c \
-                    src/allegro_4/keyboard.c \
-                    src/allegro_4/samples.c \
-                    src/allegro_4/fileutils.c \
-                    src/allegro_4/allegro_icon.c
+elkulator_hal_src = src/host_abstraction_layer/allegro_4//2xsai.c \
+                    src/host_abstraction_layer/allegro_4//adc.c  \
+                    src/host_abstraction_layer/allegro_4//palfilt.c \
+                    src/host_abstraction_layer/allegro_4//scale2x.c \
+                    src/host_abstraction_layer/allegro_4//soundopenal.c \
+                    src/host_abstraction_layer/allegro_4//firstbyte.c \
+                    src/host_abstraction_layer/allegro_4//linux-gui.c \
+                    src/host_abstraction_layer/allegro_4//linux-keydefine.c \
+                    src/host_abstraction_layer/allegro_4//video.c \
+                    src/host_abstraction_layer/allegro_4//keyboard.c \
+                    src/host_abstraction_layer/allegro_4//samples.c \
+                    src/host_abstraction_layer/allegro_4//fileutils.c \
+                    src/host_abstraction_layer/allegro_4//allegro_icon.c
 else
-elkulator_hal_src = src/allegro_5/adc.c  \
-                    src/allegro_5/event_handler.c \
-                    src/allegro_5/firstbyte.c \
-                    src/allegro_5/fileutils.c \
-                    src/allegro_5/keyboard.c \
-                    src/allegro_5/linux-keydefine.c \
-                    src/allegro_5/menus.c \
-                    src/allegro_5/menu_file.c \
-                    src/allegro_5/menu_tape.c \
-                    src/allegro_5/menu_disc.c \
-                    src/allegro_5/menu_roms.c \
-                    src/allegro_5/menu_settings.c \
-                    src/allegro_5/menu_misc.c \
-                    src/allegro_5/palfilt.c \
-                    src/allegro_5/samples.c \
-                    src/allegro_5/soundopenal.c \
-                    src/allegro_5/video.c
+elkulator_hal_src = src/host_abstraction_layer/allegro_5//adc.c  \
+                    src/host_abstraction_layer/allegro_5//event_handler.c \
+                    src/host_abstraction_layer/allegro_5//firstbyte.c \
+                    src/host_abstraction_layer/allegro_5//fileutils.c \
+                    src/host_abstraction_layer/allegro_5//keyboard.c \
+                    src/host_abstraction_layer/allegro_5//linux-keydefine.c \
+                    src/host_abstraction_layer/allegro_5//menus.c \
+                    src/host_abstraction_layer/allegro_5//menu_file.c \
+                    src/host_abstraction_layer/allegro_5//menu_tape.c \
+                    src/host_abstraction_layer/allegro_5//menu_disc.c \
+                    src/host_abstraction_layer/allegro_5//menu_roms.c \
+                    src/host_abstraction_layer/allegro_5//menu_settings.c \
+                    src/host_abstraction_layer/allegro_5//menu_misc.c \
+                    src/host_abstraction_layer/allegro_5//palfilt.c \
+                    src/host_abstraction_layer/allegro_5//samples.c \
+                    src/host_abstraction_layer/allegro_5//soundopenal.c \
+                    src/host_abstraction_layer/allegro_5//video.c
 endif
 
 elkulator_SOURCES = src/1770.c \
