+++ src/host_abstraction_layer/allegro_4/adc.c
