+++ src/host_abstraction_layer/allegro_4/allegro_icon.c
