+++ src/host_abstraction_layer/allegro_4/firstbyte.c
