+++ src/host_abstraction_layer/allegro_4/keyboard.c
@@ -1,4 +1,4 @@
-#include "common/keyboard.h"
+#include "host_abstraction_layer/keyboard.h"
 #include "elk.h"
 #include <allegro.h>
