+++ src/host_abstraction_layer/allegro_4/linux-keydefine.c
@@ -4,7 +4,7 @@
 #ifndef WIN32
 #include <allegro.h>
 #include "elk.h"
-#include "common/keyboard.h"
+#include "host_abstraction_layer/keyboard.h"
 
 int keytemp[128];
