+++ src/host_abstraction_layer/allegro_4/samples.c
@@ -1,16 +1,13 @@
-#include <allegro5/allegro.h>
-#include <allegro5/allegro_audio.h>
-#include <allegro5/allegro_acodec.h>
-#include <allegro5/allegro_primitives.h>
-#include "common/samples.h"
+#include <allegro.h>
+#include "host_abstraction_layer/samples.h"
 
-ALLEGRO_SAMPLE *seeksmp[4][2];
-ALLEGRO_SAMPLE *motorsmp[3];
-ALLEGRO_SAMPLE *tsamples[2];
+SAMPLE *seeksmp[4][2];
+SAMPLE *motorsmp[3];
+SAMPLE *tsamples[2];
 
 bool sample_seek_load(uint8_t indexA, uint8_t indexB, const char * filename) // Don't know what a and b are at the moment.
 {
-    seeksmp[indexA][indexB] = al_load_sample(filename);
+    seeksmp[indexA][indexB] = load_wav(filename);
     return (seeksmp[indexA][indexB] != NULL);
 }
 
@@ -19,7 +16,7 @@ unsigned long sample_seek_get_length(uint8_t indexA, uint8_t indexB)
     int length = 0;
     if(seeksmp[indexA][indexB])
     {
-        length = al_get_sample_length(seeksmp[indexA][indexB]);
+        length = seeksmp[indexA][indexB]->len;
     }
     return (length);
 }
@@ -29,7 +26,7 @@ void * sample_seek_get_data_ptr(uint8_t indexA, uint8_t indexB)
     void * data = 0;
     if(seeksmp[indexA][indexB])
     {
-        data = al_get_sample_data(seeksmp[indexA][indexB]);
+        data = seeksmp[indexA][indexB]->data;
     }
     return (data);
 }
@@ -39,7 +36,7 @@ int sample_seek_get_frequency(uint8_t indexA, uint8_t indexB)
     int freq = 0;
     if(seeksmp[indexA][indexB])
     {
-        freq = al_get_sample_frequency(seeksmp[indexA][indexB]);
+        freq = seeksmp[indexA][indexB]->freq;
     }
     return (freq);
 }                    
@@ -51,12 +48,12 @@ void sample_seek_destroy_all()
     {
         if (seeksmp[index][0])
         {
-            al_destroy_sample(seeksmp[index][0]);
+            destroy_sample(seeksmp[index][0]);
             seeksmp[index][0] = NULL;
         } 
         if (seeksmp[index][1]) 
         {
-            al_destroy_sample(seeksmp[index][1]);
+            destroy_sample(seeksmp[index][1]);
             seeksmp[index][1] = NULL;
         }
     }
@@ -64,7 +61,7 @@ void sample_seek_destroy_all()
 
 bool sample_motor_load(uint8_t index, const char * filename) // Don't know what a is at the moment.
 {
-    motorsmp[index] = al_load_sample(filename);
+    motorsmp[index] = load_wav(filename);
     return (motorsmp[index] != NULL);
 }
 
@@ -73,7 +70,7 @@ unsigned long sample_motor_get_length(uint8_t index)
     int length = 0;
     if(motorsmp[index])
     {
-        length = al_get_sample_length(motorsmp[index]);
+        length = motorsmp[index]->len;
     }
     return (length);
 }
@@ -83,7 +80,7 @@ void * sample_motor_get_data_ptr(uint8_t index)
     void * data = 0;
     if(motorsmp[index])
     {
-        data = al_get_sample_data(motorsmp[index]);
+        data = motorsmp[index]->data;
     }
     return (data);
 }
@@ -93,28 +90,28 @@ int sample_motor_get_frequency(uint8_t index)
     int freq = 0;
     if(motorsmp[index])
     {
-        freq = al_get_sample_frequency(motorsmp[index]);
+        freq = motorsmp[index]->freq;
     }
     return (freq);
 }    
 
 void sample_motor_destroy_all()
 {
-//    int index;
-//    for (index=0;index<3;index++)
-//    {
-//        if (motorsmp[index])
-//        {
-//            al_destroy_sample(motorsmp[index]);
-//            motorsmp[index] = NULL;
-//        } 
-//    }
+    int index;
+    for (index=0;index<3;index++)
+    {
+        if (motorsmp[index])
+        {
+            destroy_sample(motorsmp[index]);
+            motorsmp[index] = NULL;
+        } 
+    }
 }
 
 
 bool sample_tape_noise_motor_load(uint8_t index, const char * filename) // Don't know what a is at the moment.
 {
-    tsamples[index] = al_load_sample(filename);
+    tsamples[index] = load_wav(filename);
     return (tsamples[index] != NULL);
 }
 
@@ -123,7 +120,7 @@ unsigned long sample_tape_noise_get_length(uint8_t index)
     int length = 0;
     if(tsamples[index])
     {
-        length = al_get_sample_length(tsamples[index]);
+        length = tsamples[index]->len;
     }
     return (length);
 }
@@ -133,7 +130,7 @@ void * sample_tape_noise_get_data_ptr(uint8_t index)
     void * data = 0;
     if(tsamples[index])
     {
-        data = al_get_sample_data(tsamples[index]);
+        data = tsamples[index]->data;
     }
     return (data);
 }
@@ -143,7 +140,7 @@ int sample_tape_noise_get_frequency(uint8_t index)
     int freq = 0;
     if(tsamples[index])
     {
-        freq = al_get_sample_frequency(tsamples[index]);
+        freq = tsamples[index]->freq;
     }
     return (freq);
 }    
@@ -152,13 +149,13 @@ void sample_tape_noise_destroy_all()
 {
     if(tsamples[0])
     {
-        al_destroy_sample(tsamples[0]);
+        destroy_sample(tsamples[0]);
         tsamples[0] = NULL;
     }
 
     if(tsamples[1])
     {
-        al_destroy_sample(tsamples[1]);
+        destroy_sample(tsamples[1]);
         tsamples[1] = NULL;
     }
 }
