+++ src/host_abstraction_layer/allegro_4/scale2x.c
