+++ src/host_abstraction_layer/allegro_4/video_internal.h
@@ -2,7 +2,7 @@
 #define _VIDEO_INTERNAL_H
 
 #include <allegro.h>
-#include "common/video.h"
+#include "host_abstraction_layer/video.h"
 
 // Private functions to be places in hal only header.
