+++ src/host_abstraction_layer/allegro_5/event_handler.c
@@ -2,10 +2,10 @@
 
 #include <allegro5/allegro.h>
 #include <allegro5/allegro_native_dialog.h>
-#include "common/event_handler.h"
+#include "host_abstraction_layer/event_handler.h"
 #include "event_handler_internal.h"
 #include "logger.h"
-#include "common/video.h"
+#include "host_abstraction_layer/video.h"
 #include "menu_internal.h"
 #include "keyboard_internal.h"
 #include "config_vars.h"
