+++ src/host_abstraction_layer/allegro_5/fileutils.c
