+++ src/host_abstraction_layer/allegro_5/firstbyte.c
