+++ src/host_abstraction_layer/allegro_5/keyboard_internal.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 #include <allegro5/allegro.h>
-#include "common/event_handler.h"
+#include "host_abstraction_layer/event_handler.h"
 
 elk_event_t keyboard_handle_event(ALLEGRO_EVENT *event);
