+++ src/host_abstraction_layer/allegro_5/menu_file.c
@@ -15,7 +15,7 @@
 #include <allegro5/allegro_native_dialog.h>
 #include "config_vars.h"
 #include "menu_internal.h"
-#include "common/event_handler.h"
+#include "host_abstraction_layer/event_handler.h"
 #include "callback_handlers.h"
 #include "logger.h"
