+++ src/host_abstraction_layer/allegro_5/menu_misc.c
