+++ src/host_abstraction_layer/samples.h
