+++ src/host_abstraction_layer/sound.h
