+++ src/host_abstraction_layer/video.h
