+++ src/ula.c
@@ -18,8 +18,8 @@
 #include "mem.h"
 #include "ula.h"
 #include "config_vars.h"
-#include "common/sound.h"
-#include "common/video.h"
+#include "host_abstraction_layer/sound.h"
+#include "host_abstraction_layer/video.h"
 #include "logger.h"
 #include "csw.h"
 #include "uef.h"
